breaking change: change var names in network role
openwrt_network_devices:   -> openwrt_network_deviceshost
                           -> openwrt_network_devicesgroup
openwrt_network_interfaces -> openwrt_network_interfaceshost
                           -> openwrt_network_interfacesgroup

now it is possible to define interfaces and devices in the networking
role also on a host and/or group basis in ansible
It's the same as this is already been done before in the
openwrt_network_firewall role

This was necessary when having bigger environments where interfaces and
devices are the same on lots of devices. Docs will be changed, too soon.